refactoring
- component names are now more consistent
- removed old, unnecessary comments and added new ones
- HTML/CSS divs now all follow kebab-case instead of PascalCase
