+++ src/components/ItinerariesList/ItinerariesList.js
@@ -1,12 +1,7 @@
 import ItinerariesCard from "./ItinerariesCard";
 import './ItinerariesList.css'
 import React, { useState } from 'react';
-
-import {
-    Collection,
-    ScrollView,
-    Button
-  } from "@aws-amplify/ui-react";
+import { Collection, ScrollView, Button } from "@aws-amplify/ui-react";
 
 function ItinerariesList({ previews, completedPreviews, currentTripId, setCurrentTripId, resetStage }) {
     const [showCompleted, setShowCompleted] = useState(false);
@@ -29,7 +24,7 @@ function ItinerariesList({ previews, completedPreviews, currentTripId, setCurren
             columnEnd="2"
             rowStart="2"
             rowEnd="-1"
-            className="ItinerariesList"
+            className="itineraries-list"
         >
         { previews !== null ?
           <div className="list-content">
