+++ src/components/datetime/DateTime.css
@@ -51,7 +51,7 @@
     border: 1px solid var(--color-dark-gray) !important;
 }
 
-.dateTime .react-calendar__month-view__days__day--weekend {
+.date-time .react-calendar__month-view__days__day--weekend {
     color: var(--color-light-blue);
 }
 
@@ -90,7 +90,7 @@
     scale: 90%;
 }
 
-.dateTime p {
+.date-time p {
     margin-bottom: 8px;
 }
